Set resolvedCallvirt if virtual method can be resolved at compile time
Found this with CoreCLR tests. The logic that special cases the
delegates expects resolvedCallVirt to be set, but we were never setting
it for this case. This pattern was copied from CoreCLR JitInterface and
it puzzles me as to how it could possibly work there.

We ended up asking for a ReadyToRun helper to resolve the virtual call
even though the method was not virtual.