Add IEETypeNode interface
This is similar to `IMethodNode` in the sense that it lets us preserve
information about the type associated with the symbol. Same as
`IMethodNode`, there are two classes implementing this.

I pulled the decision about which one to generate into the slow path (we
no longer need to consult the compilation group in order to pick a cache
to check).