Implement CanCastTo
Type system support for answering questions about whether a type
can be cast to another type, along with unit tests.

I largely ported the C++ code from the Project N compiler to C#. The only
divergence is in the handling of casting arrays to System.Array and array
interfaces. We won't use the "treat it as `Array<T>`" wart in CoreRT.
The unit tests are all new because there were none.

I'm also adding a few convenience properties to the type system, stolen
from the Project N compiler type system:

* `Variety` - an enum on `TypeDesc` that can easily tell what kind of type it is (Array? DefType? Generic parameter?...)
* A few new `HasXXX` and `IsXXX` shortcuts on `GenericParameterDesc`. Makes code much more readable.
* `IsGenericParameter` and `IsDefType` on `TypeDesc`. This brings the existing `IsArray`, `IsPointer` etc. to completion.
* `HasVariance` on `TypeDesc`. This is a cached answer to whether a type has variance. It lets us quickly decide if we need to do the slow things.