Add support for variant interface dispatch (#1172)
This change is to emit the new rare fields within the EEType so that at
runtime, we can perform variant interface calls (example: call
`IEnumerable<object>.GetEnumerator()` on `string[]`).

The second part forces generation of methods
declaring/implementing variant interfaces. The analysis
of which ones are actually needed is rather involved
and we might choose to do that at a later time.