ComObject Cleanup Race Condition Fix
There is a race condition between apartment shutdown cleanup and the finalizer cleanup of __ComObject. The fix is to perform the cleanup only once by doing only if m_finalizer is not null.

Cleaning up the apartment stress test that found this. Cleaning up some of the PInvoke definitions and doing some more error checking.

Updating the test to use native code to create the apartment threads. This is necessary because before the test gets a chance to call CoInitialize the thread may get paused for GC where the runtime figures out the apartment of the thread(see LowLevelThread.GetCurrentApartmentType) and even though the test later calls CoInitialize and sets the apartment we continue to use the cached value. UWP apps have no supported way to create additional STA threads like this so it isn't necessary to come up with a fix in the product.

[tfs-changeset: 1599337]