Fix EEType::BaseSize computation for value types (#1204)
BaseSize should report the boxed size of the type instance for value
types.

InstanceByteCount returns the number of bytes that would be reported by
the sizeof() operator. This doesn't include the one inherited field from
System.Object. That field is present when the type is boxed and needs to
be accounted for.