Implement type casting cache
This checkin is an adaptation of a cache used in the type loader.  It is applied to our type casting APIs and brings a significant performance improvement to any type-cast-heavy workload.

The cache has an initial tuning that looks reasonable, but should still be considered provisional as we gather more data on its behavior across various workloads.

[tfs-changeset: 1600411]