Protect from arrays with ElementType.ElementSize >= 64k
This can't be encoded in the EEType (and in the array GCDescs, for that
matter).

I'm removing checked arithmetic from the GCDesc encoder because that
code runs before ComponentSize emission, but it's not the right place to
check for this, because this also applies to value types with no GC
pointers.