Remove contracts from the startup path
Seems like the reason for using `Contract.Assert` instead of
`Debug.Assert` in the startup path was that `Debug.Assert` won't work
because it's not low level enough (needs `String.Empty` to work).
`Contract.Assert` is still problematic when the condition is not met.

The right solution is to actually define a lower level Assert.