Fix incremental build (#1275)
This change fixes the incremental build. We were always rebuilding the
ExceptionHandling.S and StubDispatch.S due to incorrectly created
dependency chain. A custom target in cmake is always considered out
of date, so a combination of a custom command and a custom target
is needed to create the right dependency.