Get rid of `AddWellKnownTypes`
This method will have a tendency to become a dumping ground of
unexplained roots in the future (I just had the joy of looking through
these in the Project N dependency reducer). We should kill it before it
lays eggs.

The dependency graph should have dependencies that are traceable to the
thing that required them to become part of the graph. For the case of
`System.String` this was used as a shortcut for making string literals
work.

This will have a small perf hit, but this kind of microoptimization
should be backed by hard data quantifying how it's useful.