Updating our steam and apartment marshaling to match desktops
After un-marshaling the object reset the stream to the beginning. This addresses an issue causing the Threading\FirstParty\FTMThreadingTests to fail. What I suspect is happening although I can't verify this in the debugger is that when we create the stream we don't zero it out so it may contain contents of a previous stream that has already been released.  So if we don't reset the stream after we have un-marshaled the object and call CoReleaseMarshalData from the finalizer the stream may be positioned at another object and CoReleaseMarshalData may be releasing that object instead of the actual one that was written to the beginning of the stream.

Updating QueryInterface to call AddRef on the instance.

Updating Read and Write to use Buffer.MemoryCopy which should be faster.

Updating CopyTo to check the return value of Read and Write.

[tfs-changeset: 1607790]