Prepare EEType.cs for moving to Common, Part 1
We have like 2 managed copies of EEType readers and there is a need to read EETypes from CoreLib and CoreRT compiler tests now (as a perf optimization/maintentance improvement). Time to consolidate everything into one thing instead of creating a new thing.

EEType.cs in the type loader is in the best position to become the ultimate EEType reader, so I'm cleaning it up a little (making it use the shared enums we use in the compiler and Runtime.Base).

***Next steps after this***
1. Mechanically remove the "m_" field prefixes
2. Put all the setters under #if TYPE_LOADER_IMPLEMENTATION
3. Move the file out next to EEType.Constants.cs (mirrors it to Github)
4. Profit
5. Make Runtime.Base use the common EEType.cs and delete the inferior EEType.cs
6. Make System.Private.CoreLib use the common EEType.cs and delete RuntimeImports for all of the trivial EEType sniffing
7. Delete RuntimeExports from Runtime.Base that provide EEType sniffing
8. Write unit tests for EEType emission in the CoreRT compiler and use EEType.cs to validate the structures look as expected

[tfs-changeset: 1608978]