Delete weird GetTypeDefinition overrides (#1361)
Returning `this` (which is what the base implementation does) is as good
as the implementation I'm deleting. This operation is only meaningful
for generic type instances, but we expose the method on `TypeDesc` for
perf reasons.

The less code - the fewer bugs. Bonus points for finding the bug in the
code I'm deleting.