Fix issue 213921 - on ARM, we throw the wrong exception if we allocate an array that needs 8-byte alignment, the element count is non-negative, but so large that the array size computation overflows.
The rules are that if the element count is negative, you get an overflow exception. If the count is too large for the array to be allocated, you get an out of memory exception. This is indeed what happens with normal arrays, but with 8-byte aligned arrays we mistakenly throw an overflow exception if the element count is positive, but too large.

[tfs-changeset: 1610733]