Fix stresslog
The stresslog in corert was not working properly due to the dummy definition of
the logging macros in gcenv.base.h. I have moved those dummy macros to the
gcenv.h of the GC sample and modified gcenv.h of the runtime to include the
stressLog.h.
I also had to change two calls to Object::GetMethodTable in the logging
macros to Object::GetGCSafeMethodTable, since Object in CoreRT doesn't
contains GetMethodTable.