ICodeManager refactoring to support separated funclets
Today, we assume that funclets are separate methods, in the sense that they get unique unwind information, but that's where the separation ends.  They are still tightly coupled to their "parent method" in other ways.  One of those couplings is the fact that they must be laid out contiguously with the parent method.  These changes update the EH dispatcher and the ICodeManager abstraction to allow for funclets to be disconnected from their parent method.  This mostly falls into the category of using addresses instead of offsets when referring to handlers and when handling the special case of "remapping" a hardware fault location to a safe point in an associated handler.

[tfs-changeset: 1612757]