Stress logs not enabled by default in chk builds
The stress log is enabled by default in chk builds. In x86 this allocates 16KB per thread that is never freed which means there are only a limited number threads that can be created in a given process before the process runs out of memory. Some functional tests hit this limit.

DevDiv: 207173 tracks addressing this issue so the stress logs don't consume an unbounded amount of memory.

I am updating the runtime so that the log is not enabled by default.

[tfs-changeset: 1613363]