Delete a bunch of runtime exports that are no longer needed
EEType is now a binary contract with a shared EEType.cs reader implementation.

The various EEType probing runtime exports have been superseded by the common EEType.cs -> remove the usage of the exports and delete them.

I will do RhGetCorElementType and RhGetEETypeClassification in a separate changeset because those introduce enum types that have widespread usage and should be deleted in favor of the common ones. This change is big enough as is.

Testing: standard precheckin + NUTC 1k test suite (to test pntestcl)

[tfs-changeset: 1615277]