Delete RhGetEETypeClassification
Instead of trying to compute an enumeration, I'm switching everything to use the IsXX accessors. There's no perf advantage in computing an enum value anymore since all IsXX accessors on EETypePtr are fully inlineable with no runtime calls.

Notable thing is that I'm not introducing a replacement of the "Regular" classification. It was weird and lead to a bug around handling generic enums that I'm fixing. Instead, I'm introducing IsDefType that should do what IsDefType does in the CoreRT type system - return true if this is a struct, class, or interface (non-generic, generic definition, or instantiatied - any of them).

[tfs-changeset: 1615478]