Add compile time and runtime instrumentation support for RCWs and CCWs
This change adds compile time and runtime instrumentation support for RCWs and CCWs in order to measure code coverage. In particular, the goal of the code coverage is to compare the number of  RCWs and CCWs that MCG generate and how many of them are actually used at runtime. This change also includes a perfview extension to read the runtime data and generate report from RCW and CCW creation and QI events.

[tfs-changeset: 1616894]