Port changeset #1617250 in branche ProjNDev2 to main branch at Michal's request - here's the checkin comment for that changeset:
Fix TODO in InitializeFieldDefinition - this caused incorrect field offsets in explicit layout types.

While investigating JIT test failures (jit\SIMD\BoxUnbox.csproj and jit\SIMD\VectorMatrix.csproj) I found that the dynamic type loader computes an incorrect size for structs with explicit layout. As it turns out, the native metadata is already wrong and doesn't contain the field offset specified in the source via the FieldOffset attribute (and present in EMCA metadata).

The fix is simply to get the offset from the ECMA metadata reader. That API however returns -1 for the case where no explicit offset was specified, so that needs to be checked for.

[tfs-changeset: 1617255]