Replace use of ManagedCalloutThunk with Universal Thunk
Move to using the Universal Thunk for cases where we need to perform complex lookup to perform interface dispatch. This allows us to permit exceptions to occur in interface processing without ripping the process down. As a side benefit, we are able to reduce the amount of platform specific assembly code needed to build the product. NOTE: CoreRT still uses the old managed callout logic, as I don't have time on my schedule to write a SYS V ABI version of the universal transition thunk.

[tfs-changeset: 1618006]