Make NameMangler not depend on Compilation (#1524)
`NameMangler`'s job is to provide stable names for types and their
members. It doesn't have a connection with the compilation process and
therefore shouldn't depend on `Compilation`. This makes it reusable and
unit testable.

I'm also taking `CompilationUnitPrefix` away from it. It doesn't have a
logical connection (except that it's related to naming and it was
convenient to put it there). I'm putting it on NodeFactory temporarily,
but there is a strong correlation between a dependency node using
`CompiationUnitPrefix` in it's name and reporting
`ShouldShareNodeAcrossModules` as false. Maybe we should use that in
`ObjectWriter` to prepend it automatically instead of hoping the node
won't mess it up and jumping through hoops (a static variable) to get to
it.