Fix generated IL for the DynamicInvoke thunk
Since the HelperParamRef/In intrinsics are currently faked as being
member methods of the delegate itself (they have ByRef return values and
can't be expressed as a C# method we could put `[Intrinsic]` on), they
need to be instantiated as open before being called from the referenced
code. RyuJIT apparently doesn't care, but CppCodeGen is more sensitive.

(Boy I'm glad I spent time improving the IL disassembly -
troubleshooting this was a breeze.)