Merge TypeInitializationManager with CompilerTypeSystemContext (#1548)
There's little point in separating type initialization manager from the
type system context. It's an extra class we need to instantiate and pass
around between Compilation and NodeFactory and doesn't provide much
functionality to warrant being a separate class (we don't expect this to
be more pluggable/overridable than the type system context in general -
we might want to consider a middle man class between
MetadataTypeSystemContext and CompilerTypeSystemContext if pluggability
is desired).

Also, exposing CompilerTypeSystemContext on NodeFactory is more useful.