Drop infrastructure for never-finished ReflectionDomain feature.
Background: Many moons ago, Reflection.Core was envisioned to
be an LMR-type library that that could instantiate multiple
"type universes" (aka ReflectionDomain) along with one special
domain known as the ExecutionDomain (the one that allows invocation.)

Reality: The support for domains other than the ExecutionDomain
was never demanded or finished so in reality, ReflectionDomain
is a singleton object but we're still passing it around all over
the place and storing it in objects "just in case" we finish the feature.
There still plenty of "todo's" in the code regarding this and it'll
be an ongoing tax and bit-rot as we add to the code. Since we don't
expect to need this feature, this commit cuts out all the plumbing
and overhead (both mental and runtime) associated with it.