Adding support for call interceptors.
Call interceptors sit in between a caller and a callee adapting for different calling conventions or signatures. This will be used by interop in the future for adapting from a safe method signature to the unsafe(native) target and vice versa.

Support is currently limited to AMD64 since it only has the one calling convention.

[tfs-changeset: 1619082]