Fix bug 241854
This is to fix bug 241854, where interop generated a spin loop around a call to RhYield and, via a few levels of inlining, the compiler emitted a tight loop around a call to RhYield without loop hijacking in the loop codegen.  This is because the compiler assumes all “normal” calls (such as this one) will allow hijacking.  Unfortunately, this isn’t the case for this method, which is a RuntimeImport from System.Private.CoreLib to some runtime unmanaged code, which is not hijackable.  The fix, in this case, is to pinvoke to RhYield instead of using RuntimeImport.

[tfs-changeset: 1619386]