Add helpers to compute type hashcode from native metadata
We will be converting the type reflection mapping table (the one that
converts between EETypes and type metadata handles) to a hashtable.

In order to do that, we need to be able to compute the type hash code
from metadata. In this commit:

* I'm making it possible to compute the type hash code without
allocating and formatting the complete type name (to cut down on
allocations)
* I'm adding a helper to compute hash codes from `TypeReferenceHandle`
and `TypeDefinitionHandle`
* A bunch of unit tests for both.