Do not handle reflection blocked types as named types
This API misuse started to hit a new assertion in TryGetMetadataForNamedType that was checking that the type handle passed to it is not garbage (i.e. doesn't refer to a constructed type).

I'm moving the check into ExecutionDomain so that it's in the same spot as the other check for blocked types.

The added benefit of this is that:
* we don't decompose the generic instance type handle twice anymore (the first thing IsReflectionBlocked does with generic types is to fetch the definition)
* we don't try to find metadata for something that doesn't have it anyway because it's not a definition

[tfs-changeset: 1622087]