A small preparatory change for TypeInfo restoration.
RuntimeTypeInfo.GetDeclaredMethods() and GetDeclaredNestedType()
were inheriting their implementations from TypeInfo.
This is fine with today's bits but as soon as we replace
TypeInfo with the one from the full framework, these
will break as the full framework TypeInfo routes everything
through the BindingFlags apis which won't be implemented
yet.

To avoid that, we'll override them with a self-sufficient
implementation to tide us over until everything's
back up and running.

This also shone light on the fact that the unspoken
convention of making sure each override/no-override
decision is made explicit in this particular codebase
has bit-rotted some. So I added some new override
declarations to fill those holes.