Update metadata reader to be more lazy
Addressing a big known problem of the native metadata reader - allocating way too many arrays (and array enumerators) when reading metadata.

I'm updating the reader generator to generate XXXCollection types that are structs and have a struct enumerator.

I'm specifically not updating the consumers to take advantage of this (i.e. we still sometimes pass the collection as an IEnumerable within the reflection stack). This is because I don't want the changeset to clash with the reflection refactoring work - I would like to push this change into our next release. This makes SharedLibrary.dll 35 kB smaller on x86 because we no longer have the crazy amount of array types implementing all the crazy collection types.

[tfs-changeset: 1622617]