Another preparatory change for Reflection restoration.
If RuntimeCustomAttributeData.ToString() intercepted
a MissingMetadataException, it fell back to Object.ToString()
(via "base.ToString()").

Unfortunately, the incoming CustomAttributeData class
overrides Object.ToString() with a non-pay-for-play-safe
version of RuntimeCustomAttributeData.ToString().
(Perhaps in Project N, we'll choose to have to
thunk up to R.Core to run the pay-for-play-safe version
instead.)

Either way we decide, that would be a counterproductive
fallback, to say the least.

So rather than trusting "base", we'll directly implement
what we already know Object.ToString() does: namely,
GetType().ToString().