Fix for bug 203017: [Customer] MissingTemplateException from TypeInfo.GetCustomerAttribute<>
The Reflection Invoke logic gets tripped by the existence of enum parameters with default values in the method. NUTC generates reflection templates using the enum type , but Reflection searches for templates using the underlying type of the enum (in that case, type of int). These 2 parts of the system end up not agreeing with each other on the correct type of the parameter, and the Reflection Invoke operation ends up failing and not finding a template because of that.

[tfs-changeset: 1622806]