Modify SGen to resolve duplicate type matches through type forwards if available.
SGen will try to find types by simply looking for it in all of the loaded assemblies. This fails if we find the type twice (basically a bug in the FX, but those happen). This change modifies this code to try CCI as the last chance. We try to find a type forwarder for the affected type. If we do find it, we simply follow it, since that will pick the right assembly (and thus the right version of the type).

Also revert the hack to work around the issue which is now fixed.

[tfs-changeset: 1623908]