Bring up the apis to retrieve custom attributes.
Gee, I don't think we have enough apis to do this - should
we propose a few more?

Anyhow, this was done by moving the CustomAttributeExtensions
class from System.Reflection.Extensions.dll into Corelib.

We also needed to move CustomAttributeInstantiator
from S.P.R.Execution to Corelib (with a typeforwarder.)

Though CustomAttributeExtensions is supposed to be the
"bandaid" api, it actually has the nicest apis so now
it gets to be near the bottom of the stack (with
XInfo.CustomAttributes being the lowest-level primitive.)

Everything else is legacy api that we threw out in
.Net Core.