Fix exceptions from funclets
This change fixes exceptions from funclets - the EH info is attached to the
main function only and its child funclets didn't have any EH info reported
by the EH info enumerator.
The fix is to store reference to the main function's EH info in the child
funclets LSDA and to extract it in the UnixNativeCodeManager::FindMethodInfo.