Fix CppCodeGen build break
The last batch of changes in CoreLib exposed bug in CppCodeCode. Some of the forward type definitions were not emitted, causing build break. I am adding workaround to fix the build break.

I have also done small refactoring to eliminate unnecessary copies of larges amount of text, and fixed a few method names for consistency while I was on it.