Create a per-Type cache for BindingFlag results.
This mirrors CoreClr's strategy (organize caches
by the name used to lookup, then filter it for BindingFlags
combinations on each lookup.)

The cache is its own object so Types that don't
get queried pay the storage cost. In addition,
the creation of the cache is a useful signal
that the Type is being queried. This could be
used as a heuristic for keeping the Type alive
for longer than the normal weak-reference lifetime.

There are safeguards built in to ensure
that DeclaredOnly searches don't get bollixed due
to MissingMetadataExceptions in base classes they
have no interest in (not sure if the reducer
can generate this case today, but it's cheap
insurance.)