Do not expose Lock instances to applications
CR: KouVel, MorganB

It was possible for an application to get instances of the Lock type by using LazyInitializer:

    object syncLock = null;
    bool _ = false;
    // syncLock is assigned an instance of the Lock type
    LazyInitializer.EnsureInitialized(ref _, ref _, ref syncLock);

After change 1623787 an attempt to lock on a Lock instance causes assertion violation in Monitor.GetLock() in debug builds since we no longer have a short code path for that case.

The fix is to use "new object()" instead of "new Lock()" for creating synchronization objects in LazyInitializer.

[tfs-changeset: 1628141]