Replace all uses of FoundationTypes and typeof()
... in Reflection.Core with CommonRuntimeTypes.X
where possible. FoundationTypes is a relic from
a time where Reflection.Core was supposed to handle
multiple type universes. typeof() for commonly used
type still incurs the overhead of a hash lookup at
best (and the recreation of the Type object at worst.)

Holding off on getting rid of FoundationTypes
itself for now to avoid colliding with the
AssemblyBinder migration.