Get rid of useless AsType()/GetTypeInfo() calls.
We no longer need most of those now that we've backtracked
out of that api refactoring and they're just wasting
cycles.

And yes, I realize that a lot of the variables
in Assignability.cs are still named "TypeInfo".
Maybe in a future cleanup, but right now, leaving
those names as is makes this commit easier to diff.

The new GetGenericTypeParameters() extension method
gives us the moral equivalent of TypeInfo.GenericTypeParameters
on Type. (And yes, I know we could just call Type.GetGenericArguments()
but I detest that semantic overloading.)