Make GetName protected internal
`GetName` is "get the name to be used in the diagnostic log". The rules
around it in the compier are:

* if the node is a `ISymbolNode` it should include the mangled name so
that it's easy to crossreference it with the object file (doesn't
necessarily have to be equal to the mangled name - see EETypeNode and
friends that append "constructed", etc.)
* if the node is not an `ISymbolNode`, get something that will be useful
to search for in the dependency log

I'm making the API protected and internal to avoid misuse and fixing the
places misusing this API (the spot in ObjectWriter was just wrong,
CppWriter was okay, but with problems waiting to happen).

I'm also unsealing ToString. We'll want to override it on a case-by-case
basis to make debugger experience more pleasant.