Add Test.CoreLib
This is a minimal core library for test purposes. It doesn't have
support for a broad range of features that people are used to (proper
class constructor support, reflection, glob/loc, etc), but on the other
hand lets you exercise various compiler features that would otherwise
not be very testable (because unless the feature is already fully
implemented, you won't be able to run enough code to get to your test).

It's largely based on stealing files from Runtime.Base and I think I
found a sustainable way to extend it as needed (with... drumroll...
partial classes).